Move State in dedicated folder to avoid circular dependencies
