+++ src/google/adk/sessions/in_memory_session_service.py
@@ -21,12 +21,12 @@ import uuid
 from typing_extensions import override
 
 from ..events.event import Event
+from ..states import State
 from .base_session_service import BaseSessionService
 from .base_session_service import GetSessionConfig
 from .base_session_service import ListEventsResponse
 from .base_session_service import ListSessionsResponse
 from .session import Session
-from .state import State
 
 
 class InMemorySessionService(BaseSessionService):
