+++ tests/unittests/tools/openapi_tool/openapi_spec_parser/test_rest_api_tool.py
@@ -20,7 +20,7 @@ from fastapi.openapi.models import MediaType, Operation
 from fastapi.openapi.models import Parameter as OpenAPIParameter
 from fastapi.openapi.models import RequestBody
 from fastapi.openapi.models import Schema as OpenAPISchema
-from google.adk.sessions.state import State
+from google.adk.states.state import State
 from google.adk.tools.openapi_tool.auth.auth_helpers import token_to_scheme_credential
 from google.adk.tools.openapi_tool.common.common import ApiParameter
 from google.adk.tools.openapi_tool.openapi_spec_parser.openapi_spec_parser import OperationEndpoint
